cleanup project files
